deprecate the filter types and move to catalog-node
