+++ plugins/catalog-backend/src/alpha.ts
@@ -14,11 +14,5 @@
  * limitations under the License.
  */
 
-// TODO(Rugvip): Re-exported for alpha types as the API report will otherwise
-//               produce warnings due to the indirect dependency. Would be nice to avoid.
-import type { EntitiesSearchFilter } from './catalog/types';
-
-export type { /** @alpha */ EntitiesSearchFilter };
-
 export * from './permissions';
 export { catalogPlugin as default } from './service/CatalogPlugin';
