+++ plugins/catalog-backend/src/permissions/rules/isEntityKind.ts
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 import { RESOURCE_TYPE_CATALOG_ENTITY } from '@backstage/plugin-catalog-common/alpha';
 import { z } from 'zod';
-import { EntitiesSearchFilter } from '../../catalog/types';
 import { createCatalogPermissionRule } from './util';
 
 /**
@@ -36,7 +36,7 @@ export const isEntityKind = createCatalogPermissionRule({
     const resourceKind = resource.kind.toLocaleLowerCase('en-US');
     return kinds.some(kind => kind.toLocaleLowerCase('en-US') === resourceKind);
   },
-  toQuery({ kinds }): EntitiesSearchFilter {
+  toQuery({ kinds }) {
     return {
       key: 'kind',
       values: kinds.map(kind => kind.toLocaleLowerCase('en-US')),
