+++ plugins/catalog-backend/src/service/AuthorizedEntitiesCatalog.test.ts
@@ -20,8 +20,9 @@ import { createConditionTransformer } from '@backstage/plugin-permission-node';
 import { isEntityKind } from '../permissions/rules/isEntityKind';
 import { CatalogPermissionRule } from '../permissions/rules';
 import { AuthorizedEntitiesCatalog } from './AuthorizedEntitiesCatalog';
-import { Cursor, EntityFilter, QueryEntitiesResponse } from '../catalog/types';
+import { Cursor, QueryEntitiesResponse } from '../catalog/types';
 import { Entity } from '@backstage/catalog-model';
+import { EntityFilter } from '@backstage/plugin-catalog-node';
 
 describe('AuthorizedEntitiesCatalog', () => {
   const fakeCatalog = {
