+++ plugins/catalog-backend/src/service/AuthorizedEntitiesCatalog.ts
@@ -35,12 +35,12 @@ import {
   EntityAncestryResponse,
   EntityFacetsRequest,
   EntityFacetsResponse,
-  EntityFilter,
   QueryEntitiesRequest,
   QueryEntitiesResponse,
 } from '../catalog/types';
 import { basicEntityFilter } from './request/basicEntityFilter';
 import { isQueryEntitiesCursorRequest } from './util';
+import { EntityFilter } from '@backstage/plugin-catalog-node';
 
 export class AuthorizedEntitiesCatalog implements EntitiesCatalog {
   constructor(
