+++ plugins/catalog-node/src/alpha.ts
@@ -19,9 +19,6 @@ export type { CatalogProcessingExtensionPoint } from './extensions';
 export { catalogProcessingExtensionPoint } from './extensions';
 export type { CatalogAnalysisExtensionPoint } from './extensions';
 export { catalogAnalysisExtensionPoint } from './extensions';
-export type {
-  EntitiesSearchFilter,
-  CatalogPermissionRuleInput,
-} from './extensions';
+export type { CatalogPermissionRuleInput } from './extensions';
 export type { CatalogPermissionExtensionPoint } from './extensions';
 export { catalogPermissionExtensionPoint } from './extensions';
