+++ plugins/catalog-node/src/api/common.ts
@@ -52,3 +52,36 @@ export type EntityRelationSpec = {
    */
   target: CompoundEntityRef;
 };
+
+/**
+ * A filter expression for entities.
+ *
+ * @public
+ */
+export type EntityFilter =
+  | { allOf: EntityFilter[] }
+  | { anyOf: EntityFilter[] }
+  | { not: EntityFilter }
+  | EntitiesSearchFilter;
+
+/**
+ * Matches rows in the search table.
+ *
+ * @public
+ */
+export type EntitiesSearchFilter = {
+  /**
+   * The key to match on.
+   *
+   * Matches are always case insensitive.
+   */
+  key: string;
+
+  /**
+   * Match on plain equality of values.
+   *
+   * Match on values that are equal to any of the given array items. Matches are
+   * always case insensitive.
+   */
+  values?: string[];
+};
