+++ plugins/catalog-node/src/api/index.ts
@@ -15,7 +15,12 @@
  */
 
 export { processingResult } from './processingResult';
-export type { EntityRelationSpec, LocationSpec } from './common';
+export type {
+  EntityRelationSpec,
+  LocationSpec,
+  EntitiesSearchFilter,
+  EntityFilter,
+} from './common';
 export type {
   CatalogProcessor,
   CatalogProcessorParser,
