+++ plugins/catalog-node/src/extensions.ts
@@ -13,14 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 import { createExtensionPoint } from '@backstage/backend-plugin-api';
+import { Entity } from '@backstage/catalog-model';
 import {
-  EntityProvider,
   CatalogProcessor,
+  EntitiesSearchFilter,
+  EntityProvider,
   PlaceholderResolver,
   ScmLocationAnalyzer,
 } from '@backstage/plugin-catalog-node';
-import { Entity } from '@backstage/catalog-model';
 import { PermissionRuleParams } from '@backstage/plugin-permission-common';
 import { PermissionRule } from '@backstage/plugin-permission-node';
 
@@ -59,14 +61,6 @@ export const catalogAnalysisExtensionPoint =
   createExtensionPoint<CatalogAnalysisExtensionPoint>({
     id: 'catalog.analysis',
   });
-
-/**
- * @alpha
- */
-export type EntitiesSearchFilter = {
-  key: string;
-  values?: string[];
-};
 
 /**
  * @alpha
